+++ forui/lib/src/widgets/breadcrumb.dart
@@ -144,7 +144,7 @@ class _Crumb extends StatelessWidget implements FBreadcrumbItem {
           (_, data, child) => Padding(
             padding: style.padding,
             child: DefaultTextStyle(
-              style: switch ((current, data.hovered)) {
+              style: switch ((current, data.hovered || data.pressed)) {
                 (false, false) => style.unselectedTextStyle,
                 (false, true) => style.hoveredTextStyle,
                 (true, true) => style.hoveredTextStyle,